Rename presentation layer files
