+++ app/src/main/java/ru/neexol/rtut/presentation/screens/map/MapViewModel.kt
@@ -1,4 +1,4 @@
-package ru.neexol.rtut.presentation.screens.maps
+package ru.neexol.rtut.presentation.screens.map
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -7,16 +7,15 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import ru.neexol.rtut.data.maps.MapsRepository
 import javax.inject.Inject
 
 @HiltViewModel
-class MapsViewModel @Inject constructor(
+class MapViewModel @Inject constructor(
 	private val repo: MapsRepository
 ) : ViewModel() {
-	var uiState by mutableStateOf(MapsUiState())
+	var uiState by mutableStateOf(MapUiState())
 		private set
 
 	var classroom by mutableStateOf("")
@@ -27,7 +26,7 @@ class MapsViewModel @Inject constructor(
 		fetchJob = viewModelScope.launch {
 			repo.getMaps(classroom).collect { maps ->
 				uiState = maps.to(
-					onSuccess = { MapsUiState(maps = it.maps, floor = it.floor, classroom = it.classroom) },
+					onSuccess = { MapUiState(maps = it.maps, floor = it.floor, classroom = it.classroom) },
 					onFailure = { uiState.copy(isMapsLoading = false, message = it.toString()) },
 					onLoading = { uiState.copy(isMapsLoading = true) }
 				)
