+++ app/src/main/java/ru/neexol/rtut/presentation/screens/schedule/ScheduleUiState.kt
@@ -1,9 +1,9 @@
-package ru.neexol.rtut.presentation.screens.grouplessons
+package ru.neexol.rtut.presentation.screens.schedule
 
 import ru.neexol.rtut.data.lessons.models.Lesson
 import ru.neexol.rtut.data.lessons.models.LessonTime
 
-data class GroupLessonsUiState(
+data class ScheduleUiState(
 	val lessons: List<List<List<Lesson?>>>? = null,
 	val isLessonsLoading: Boolean = false,
 	val times: List<LessonTime>? = null,
