+++ app/src/main/java/ru/neexol/rtut/presentation/screens/teacher/TeacherUiState.kt
@@ -1,9 +1,9 @@
-package ru.neexol.rtut.presentation.screens.teacherlessons
+package ru.neexol.rtut.presentation.screens.teacher
 
 import ru.neexol.rtut.data.lessons.models.Lesson
 import ru.neexol.rtut.data.lessons.models.LessonTime
 
-data class TeacherLessonsUiState(
+data class TeacherUiState(
 	val lessons: List<List<List<Lesson>>>? = null,
 	val isLessonsLoading: Boolean = false,
 	val times: List<LessonTime>? = null,
